+++ CHANGELOG.md
@@ -20,11 +20,19 @@ Django upgrade includes database changes, run migrations
 - Requirements version updates
 - Added missing text field to partial result nested serializer
 
+## 0.5.0 - 2020-09-13
+### Changes
+- Added statistics links
+
 ## 0.4.0 - 2020-08-14
 ### Changes
 - Added group by one support to statistics list
 - Fixed missing user id in logging
 - Changed parameter to limit organizations to users own 
+
+### Updating notes
+Includes database changes, run migrations
+- Added StatisticsLink model
 
 ## 0.3.0 - 2020-08-11
 ### Changes